[NEW] simplify bot.js file
