+++ src/bot.js
@@ -13,17 +13,15 @@ import { join } from "path";
 import { Low, JSONFile } from "lowdb";
 
 // commands imports
-import commands from "./commands/index.js";
 import { wishBirthday } from "./commands/birthday.js";
 import { PERSONALITY } from "./commands/personality.js";
 // helpers imports
 import {
-  isAdmin,
-  isCommand,
-  reactionHandler,
-  checkIsOnThread,
   generateSpotifyClient,
   removeSpotify,
+  onPrivateMessage,
+  onPublicMessage,
+  removeReminder,
 } from "./helpers/index.js";
 // jsons imports
 import commons from "./jsons/commons.json";
@@ -105,105 +103,26 @@ const self = process.env.CLIENTID; // get self Discord Id
 // Bot event FUNCTIONS
 const onMessageHandler = async (message) => {
   // Function triggered for each message sent
-  const { channel, author, content } = message;
+  const { channel } = message;
 
   if (channel.type === "DM") {
-    onPrivateMessage(message);
+    onPrivateMessage(message, client);
   } else {
     const currentServer = commons.find(
       ({ guildId }) => guildId === channel.guild.id
     );
-
-    if (
-      author.id === self || // ignoring message from himself
-      !currentServer || // ignoring if wrong guild
-      (process.env.DEBUG === "yes" && currentServer.name === "prod") // ignoring if debug && prod
-    )
-      return;
-
-    const { playlistThreadId } = currentServer;
-
-    reactionHandler(message, content, currentServer, client);
-
-    // check for command
-    const commandName = content.split(" ")[0];
-    const command = commands
-      .filter(({ admin }) => (admin && isAdmin(author.id)) || !admin) //filter appropriate commands if user has or not admin rigths
-      .find(({ name }) => commandName.slice(1) === name);
-    if (command && isCommand(commandName)) {
-      if (
-        command.name === "spotify" &&
-        process.env.USE_SPOTIFY === "yes" &&
-        channel.id === playlistThreadId
-      ) {
-        // spotify stuff
-        checkIsOnThread(channel, playlistThreadId); //add bot if not on thread
-      }
-      command.action(message, PERSONALITY.commands, client, currentServer);
-    }
+    onPublicMessage(message, client, currentServer, self);
   }
 };
 
 const onReactionHandler = async (messageReaction) => {
-  const { message, emoji, users } = messageReaction;
   const currentServer = commons.find(
-    ({ guildId }) => guildId === message.channel.guild.id
+    ({ guildId }) => guildId === messageReaction.message.channel.guild.id
   );
 
-  const { removeEmoji } = currentServer;
   removeSpotify(messageReaction, PERSONALITY, client, currentServer);
 
-  const foundReminder = client.remindme.find(
-    // found corresponding reminder message
-    ({ botMessage }) => botMessage.id === message.id
-  );
-  if (
-    foundReminder &&
-    emoji.name === removeEmoji &&
-    users.cache // if user reacting is the owner of reminder
-      .map((user) => user.id)
-      .includes(message.mentions.users.first().id)
-  ) {
-    try {
-      console.log("coucou");
-      client.remindme = client.remindme.filter(({ botMessage, timeout }) => {
-        if (botMessage.id === message.id) {
-          // if it is the right message
-          console.log("salut");
-          clearTimeout(timeout); //cancel timeout
-          botMessage.reply(PERSONALITY.commands.reminder.delete);
-          return false;
-        }
-        return true;
-      });
-      return;
-    } catch (err) {
-      console.log("reminderError", err);
-    }
-  }
-};
-
-const onPrivateMessage = async (message) => {
-  const { author, content } = message;
-
-  if (!isAdmin(author.id)) return; // If not admin, no rigth to
-
-  const destinationChannelId = content.split(" ")[0];
-
-  const newContent = content.split(" ").slice(1).join(" ");
-
-  try {
-    const channel = await client.channels.fetch(destinationChannelId);
-
-    if (channel) {
-      channel.sendTyping(); // effect of Ewibot writing
-      setTimeout(() => {
-        channel.send(newContent);
-      }, 2000); // duration
-    }
-  } catch (e) {
-    console.log(e);
-  }
+  removeReminder(messageReaction, client, currentServer);
 };
 
 // Create an event listener for messages
