+++ src/helpers/listeners.js
@@ -1,0 +1,97 @@
+import commands from "../commands/index.js";
+import {
+  isAdmin,
+  isCommand,
+  reactionHandler,
+  checkIsOnThread,
+} from "./index.js";
+
+import { PERSONALITY } from "../commands/personality.js";
+
+export const onPrivateMessage = async (message, client) => {
+  const { author, content } = message;
+
+  if (!isAdmin(author.id)) return; // If not admin, no rigth to
+
+  const destinationChannelId = content.split(" ")[0];
+
+  const newContent = content.split(" ").slice(1).join(" ");
+
+  try {
+    const channel = await client.channels.fetch(destinationChannelId);
+
+    if (channel) {
+      channel.sendTyping(); // effect of Ewibot writing
+      setTimeout(() => {
+        channel.send(newContent);
+      }, 2000); // duration
+    }
+  } catch (e) {
+    console.log(e);
+  }
+};
+
+export const onPublicMessage = (message, client, currentServer, self) => {
+  const { author, content, channel } = message;
+  if (
+    author.id === self || // ignoring message from himself
+    !currentServer || // ignoring if wrong guild
+    (process.env.DEBUG === "yes" && currentServer.name === "prod") // ignoring if debug && prod
+  )
+    return;
+
+  const { playlistThreadId } = currentServer;
+
+  reactionHandler(message, content, currentServer, client);
+
+  // check for command
+  const commandName = content.split(" ")[0];
+  const command = commands
+    .filter(({ admin }) => (admin && isAdmin(author.id)) || !admin) //filter appropriate commands if user has or not admin rigths
+    .find(({ name }) => commandName.slice(1) === name);
+  if (command && isCommand(commandName)) {
+    if (
+      command.name === "spotify" &&
+      process.env.USE_SPOTIFY === "yes" &&
+      channel.id === playlistThreadId
+    ) {
+      // spotify stuff
+      checkIsOnThread(channel, playlistThreadId); //add bot if not on thread
+    }
+    command.action(message, PERSONALITY.commands, client, currentServer);
+  }
+};
+
+export const removeReminder = (messageReaction, client, currentServer) => {
+  const { removeEmoji } = currentServer;
+  const { message, emoji, users } = messageReaction;
+
+  const foundReminder = client.remindme.find(
+    // found corresponding reminder message
+    ({ botMessage }) => botMessage.id === message.id
+  );
+  if (
+    foundReminder &&
+    emoji.name === removeEmoji &&
+    users.cache // if user reacting is the owner of reminder
+      .map((user) => user.id)
+      .includes(message.mentions.users.first().id)
+  ) {
+    try {
+      console.log("coucou");
+      client.remindme = client.remindme.filter(({ botMessage, timeout }) => {
+        if (botMessage.id === message.id) {
+          // if it is the right message
+          console.log("salut");
+          clearTimeout(timeout); //cancel timeout
+          botMessage.reply(PERSONALITY.commands.reminder.delete);
+          return false;
+        }
+        return true;
+      });
+      return;
+    } catch (err) {
+      console.log("reminderError", err);
+    }
+  }
+};
